pass int to set_num_threads()

+++ llm_bench/python/benchmark.py
@@ -742,7 +742,7 @@ def main():
         if model_args['config'].get('PREC_BF16') and model_args['config']['PREC_BF16'] is True:
             log.warning('[Warning] Param bf16/prec_bf16 only work for framework pt. It will be disabled.')
         if model_args['num_beams'] > 1:
-            torch.set_num_threads(torch.get_num_threads() / 2)
+            torch.set_num_threads(int(torch.get_num_threads() / 2))
         else:
             torch.set_num_threads(1)
     log.info(out_str)
